requirements.txt
langchain_openai
langgraph
pydantic
python-dotenv
openai
bs4
requests
matplotlib
wordcloud
textblob
